refactored all python code structure below a common 'challenge_fund_db' package name
