+++ python/challenge_fund_db/common/base.py
