+++ python/challenge_fund_db/loss/export_loss_model.py
@@ -28,8 +28,8 @@ import datetime
 
 from loss_model import LossModel, LossMap, LossMapValue, \
     LossCurveMap, LossCurveMapValue
-from cf_common import Contribution
-from database import db_connections
+from challenge_fund_db.common import Contribution
+from challenge_fund_db.common import db_connections
 import db_settings
 
 VERBOSE = True
